# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-spheresurfacearea",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — compute the surface area of a sphere, `A = 4 * pi * r^2`.",
        "description": "This service owns the *control flow* but delegates every arithmetic step to\nits dependencies, exactly as specified:\n\n1. ask `srvcs-pi` (with an empty body, since it is a constant service) for\n   `p = pi`;\n2. ask `srvcs-floatmultiply` for `r2 = radius * radius`;\n3. ask `srvcs-floatmultiply` for `t = p * r2`;\n4. ask `srvcs-floatmultiply` for `result = 4 * t`.\n\nThis service never validates `radius` itself; if a dependency rejects the\ninput it forwards the `422`, and if a dependency is unreachable it reports\nitself degraded (`503`).",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SphereSurfaceAreaResponse"
                }
              }
            }
          },
          "422": {
            "description": "a dependency rejected the input (forwarded)"
          },
          "500": {
            "description": "a dependency returned a malformed result"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "radius"
        ],
        "properties": {
          "radius": {
            "type": "object",
            "description": "The radius of the sphere."
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      },
      "SphereSurfaceAreaResponse": {
        "type": "object",
        "required": [
          "radius",
          "result"
        ],
        "properties": {
          "radius": {
            "type": "object"
          },
          "result": {
            "type": "number",
            "format": "double"
          }
        }
      }
    }
  }
}